rename package pkg/httpserver to linksharing/httpserver
