+++ cmd/linksharing/main.go
@@ -16,8 +16,8 @@ import (
 	"storj.io/storj/internal/fpath"
 	"storj.io/storj/lib/uplink"
 	"storj.io/storj/linksharing"
+	"storj.io/storj/linksharing/httpserver"
 	"storj.io/storj/pkg/cfgstruct"
-	"storj.io/storj/pkg/httpserver"
 	"storj.io/storj/pkg/process"
 )
